Added a stage to perform SonarQube analysis
This stage uses a custom image (which resides inside jenkins-build-envs)
that creates a docker container with sonar-scanner already installed.

This allows Sonar Scanner CLI (/usr/bin/sonar-scanner) to be used to
peform analysis on a project and is language and build tool agnostic

The sonar analysis phase depends on the reports from unit-test and other
static analysis phases being stashed so that they can be used in the
sonar stage.

The pylint command has been amended to:
    1. Display report on console (to be used by the pylint plugin and
display reports directly on the Jenkins build page) and,
    2. Redirect to a file which is stashed and used by the sonar stage
    3. finally, the actual format had to be amended to the one mandated
by the SonarPython plugin for pylint reports.