Specify Sonar HOST using a configuration file
The sonar host was being specified in
es-python-poc/sonar-project.properties and since that is considered
public, internal URLs should be redacted from it.

This configuration is now being read from a config file created in
Jenkins under a specific folder. This also means, that only jobs within
this folder will be able to source this configuration file and be able
to successfully access the SonarQube server.